Split HealthParser data access out of DatabaseService.SystemEvents.cs
DatabaseService.SystemEvents.cs mixed two concerns: default-trace / config-change
auditing and a ~750-line block of system_health HealthParser queries. Moved the
eight GetHealthParser*Async methods into a new DatabaseService.HealthParser.cs
partial (same class, same namespace), leaving the trace/config-change and
current-config methods in SystemEvents.

Pure partial-class split — behavior-neutral, no call-site changes. SystemEvents
1404 -> 655; HealthParser.cs 771. Dashboard-only; unrelated to the alert/tray code.
Solution builds clean; Dashboard.Tests (488) pass.
